Model main input and button with HTML form element
The "on:change" binding was in place to allow submission via the enter
key but it triggers in other unfavorable situations like when the input
looses focus leaving the user with unclear feedback. With a form now the
enter key triggers the submit event and the button becomes of type
submit by default allowing it to trigger submission as well avoiding the
unexpected behavior previously present.