Add button to search for recipes
Presumably the user intends to find a meal to eat/cook. Offering a
search for a recipe as the main call-to-action on the site should
support this goal and hopefully weaken the tendency to use this as
a 'slot machine', i.e. just gambling from one suggestion to the
next.